otk: merge directive.py into transform.py
Some of the functions in directive.py would will require resolving
values recursively.  For example, include() should run the loaded tree
through the top-level resolve() function to resolve the whole content of
the included file recursively.

Merge the two modules so that we don't end up with circular imports.
